+++ src/app/modules/user/user.service.ts
@@ -1,3 +1,5 @@
+import { StatusCodes } from 'http-status-codes';
+import AppError from '../../helpers/AppError';
 import Blog from '../blog/blog.model';
 import { IUser } from './user.interface';
 import User from './user.model';
@@ -12,7 +14,8 @@ const createUser = async (payload: IUser): Promise<IUser> => {
 const updateUser = async (userId: string, data: IUser) => {
   const user = await User.findOne({ _id: userId });
   if (!user) {
-    throw new Error('This user is not exist');
+    // throw new Error('This user is not exists');
+    throw new AppError(StatusCodes.BAD_REQUEST, 'This user is not exists');
   }
 
   if (user.role === 'admin') {
@@ -25,6 +28,11 @@ const updateUser = async (userId: string, data: IUser) => {
   return result;
 };
 const deleteBlogToDB = async (id: string) => {
+  const blog = await Blog.findOne({ _id: id });
+  if (!blog) {
+    throw new AppError(StatusCodes.BAD_REQUEST, 'This blog is not exists');
+  }
+
   const result = await Blog.findByIdAndDelete(id);
 
   return result;
